rename : 유틸리티 폴더 하위 폴더 생성 및 파일 폴더 분류 변경
